+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/referralmanagement/beneficiarydownsync/Downsync.java
@@ -16,6 +16,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.egov.common.models.service.Service;
 
 @Data
 @NoArgsConstructor
@@ -42,7 +43,10 @@ public class Downsync {
 	private List<SideEffect> SideEffects;
 	
     @JsonProperty("Referrals")
-	private List<Referral> Referrals;	
+	private List<Referral> Referrals;
+
+    @JsonProperty("Services")
+    private List<Service> services;
     
     @JsonProperty("DownsyncCriteria")
     private DownsyncCriteria downsyncCriteria;