+++ health-services/libraries/health-services-models/CHANGELOG.md
@@ -1,5 +1,9 @@
 All notable changes to this module will be documented in this file.
 
+## 1.0.26 - 2025-04-08
+- Added Service request models
+- Added Relationship model and added for HouseholdMember
+
 ## 1.0.25 - 2025-01-03
 - Added BeneficiaryType Enum and update in ProjectType, Target models
 - Added HouseHoldType Enum and added in Household, HouseholdSearch models